Refactor: consolidate reporting API endpoint into recipes module
